reorganize subject files (#801)
