+++ opencompass/datasets/subjective/__init__.py
@@ -1,0 +1,6 @@
+from .alignbench import AlignmentBenchDataset  # noqa: F401, F403
+from .corev2 import Corev2Dataset  # noqa: F401, F403
+from .creationbench import CreationBenchDataset  # noqa: F401, F403
+from .information_retrival import IRDataset  # noqa: F401, F403
+from .multiround import MultiroundDataset  # noqa: F401, F403
+from .subjective_cmp import SubjectiveCmpDataset  # noqa: F401, F403
