+++ opencompass/datasets/subjective/subjective_cmp.py
@@ -5,7 +5,7 @@ from datasets import Dataset, DatasetDict
 
 from opencompass.registry import LOAD_DATASET
 
-from .base import BaseDataset
+from ..base import BaseDataset
 
 
 @LOAD_DATASET.register_module()
