+++ opencompass/summarizers/__init__.py
@@ -1,10 +1,4 @@
 # flake8: noqa: F401, E501
-from .alignmentbench import AlignmentBenchSummarizer
 from .circular import CircularSummarizer  # noqa: F401
-from .corev2 import Corev2Summarizer  # noqa: F401
-from .creationbench import CreationBenchSummarizer
-from .creationv01 import Creationv01Summarizer  # noqa: F401
 from .default import DefaultSummarizer  # noqa: F401
-from .information_retrival import IRSummarizer  # noqa: F401
-from .multiround import MultiroundSummarizer  # noqa: F401
-from .subjective import SubjectiveSummarizer  # noqa: F401
+from .subjective import *  # noqa: F401
