+++ opencompass/summarizers/subjective/corev2.py
@@ -124,7 +124,7 @@ class Corev2Summarizer:
                         print('There are no results for ' + filename + ' or ' +
                               partial_filename)
                         print('*' * 100)
-                        assert len(result > 0)
+                        assert len(result) > 0
 
                     judged_answers = []
                     references = []
